% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{icebergr_create_namespace}
\alias{icebergr_create_namespace}
\title{Create a namespace}
\usage{
icebergr_create_namespace(catalog, namespace)
}
\arguments{
\item{catalog}{An \code{icebergr_catalog} from \code{\link[=icebergr_catalog]{icebergr_catalog()}}.}

\item{namespace}{The namespace to create. Accepts \code{"db"} or \code{c("a", "b")}.}
}
\value{
\code{catalog}, invisibly.
}
\description{
Create a namespace
}
\examples{
warehouse <- tempfile("warehouse")
dir.create(warehouse)
catalog <- icebergr_catalog("memory", warehouse = warehouse)
icebergr_create_namespace(catalog, "db")
icebergr_list_namespaces(catalog)
}
